Folder cleanup and restructure. Added Dockerfile, dockerignore, requirements.txt. Modified and renamed index function to home. Will now read from index.html file and display "Project Purple Cow". Tested and working.
