+++ core/models.py
@@ -1,0 +1,4 @@
+from django.db import models
+
+class Item(models.Model):
+    name = models.CharField(max_length=100)
