+++ core/templates/index.html
@@ -1,0 +1,1 @@
+Project Purple Cow
