--- pyapi/myapp/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MyappConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'myapp'
