--- pyapi/myapp/models.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-
-class Item(models.Model):
-    name = models.CharField(max_length=100)
