+++ requirements.txt
@@ -1,0 +1,4 @@
+asgiref==3.5.1
+Django==4.0.4
+sqlparse==0.4.2
+tzdata==2022.1
